Refactor command parsing
