+++ src/main/java/moe/das/dasshop/commands/internal/Command.java
@@ -1,6 +1,9 @@
 package moe.das.dasshop.commands.internal;
 
-import org.bukkit.command.CommandSender;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public abstract class Command {
     private final String name;
@@ -9,7 +12,13 @@ public abstract class Command {
         this.name = name;
     }
 
-    public String getName() {
+    public final String getName() {
         return name;
+    }
+
+    public final Set<Method> getSubcommandMethods() {
+        return Arrays.stream(getClass().getMethods())
+                .filter(method -> method.isAnnotationPresent(SubCommand.class))
+                .collect(Collectors.toSet());
     }
 }
